requirements.txt
fastapi
uvicorn
pdfplumber
python-dotenv
python-multipart
pymongo==3.11
dnspython


